+++ Guilherme_Ricardo_Dos_Santos_Perinotti/segundo bimestre/recharge-api/src/main/java/br/com/fag/infra/panache/repository/PanacheDataBaseRepository.java
@@ -1,4 +1,4 @@
-package br.com.fag.infra.panache;
+package br.com.fag.infra.panache.repository;
 
 import br.com.fag.domain.entities.RechargeBO;
 import br.com.fag.domain.repositories.IRechargeDataBaseRepository;
@@ -10,13 +10,13 @@ import jakarta.enterprise.context.ApplicationScoped;
 @ApplicationScoped
 public class PanacheDataBaseRepository implements IRechargeDataBaseRepository {
 
-    @Override
-    public RechargeBO persist(RechargeBO bo) {
-        PanacheRecharge entity = PanacheRechargeMapper.toEntity(bo);
+  @Override
+  public RechargeBO persist(RechargeBO bo) {
+    PanacheRecharge entity = PanacheRechargeMapper.toEntity(bo);
 
-        entity.persistAndFlush();
+    entity.persistAndFlush();
 
-        return PanacheRechargeMapper.toDomain(entity);
-    }
+    return PanacheRechargeMapper.toDomain(entity);
+  }
 
 }
